requirements.txt
fastapi
uvicorn
google-genai
chromadb
pydantic
watchfiles
pypdf